+++ tests/supportingResources/variables.tf
@@ -1,0 +1,25 @@
+variable "cidr_block" {
+  type        = string
+  description = "(Optional) The IPv4 CIDR block for the VPC. CIDR can be explicitly set or it can be derived from IPAM using `ipv4_netmask_length`."
+  default     = "10.1.0.0/16"
+}
+
+variable "name" {
+  type        = string
+  description = "Input the name of vpc stack"
+  default     = "terraform-aws-ecs-cluster"
+}
+
+variable "tags" {
+  description = " A map of tags to assign to the object. If configured with a provider default_tags configuration block present, tags with matching keys will overwrite those defined at the provider-level."
+  type        = map(string)
+  default = {
+    Environment        = "example"
+    "user::CostCenter" = "terraform-registry"
+    Department         = "DevOps"
+    Project            = "Examples"
+    Owner              = "Boldlink"
+    LayerName          = "Example"
+    LayerId            = "Example"
+  }
+}